Update `ejs` to v2.2.4
The update to the v2 branch of `ejs` causes a trailing line break to be
removed from the end of a file.

This PR adds back in that line break so as to maintain a consistent
api with v1 of `gulp-ejs`.

There is currently an open issue in mde/ejs#60 that addresses the
issue of the removal of the character.

This PR can remain unmerged until mde/ejs#60 is resolved.